# docs/reused-openapi/openapi.yaml
openapi: 3.0.3
info:
  title: EV Charging Station API
  description: |
    API για τη διαχείριση σταθμών φόρτισης ηλεκτρικών οχημάτων (EV Charging Stations).
    
    ## Αυθεντικοποίηση
    Η πλειοψηφία των endpoints απαιτεί αυθεντικοποίηση μέσω JWT token.
    Το token λαμβάνεται από το endpoint `/api/v1/auth/signin` και πρέπει να 
    συμπεριλαμβάνεται στο header `Authorization` ως `Bearer <token>`.
    
    ## Ρόλοι Χρηστών
    - **USER**: Βασικός χρήστης - μπορεί να δει φορτιστές, να κάνει κρατήσεις, κ.λπ.
    - **ADMIN**: Διαχειριστής - πλήρης πρόσβαση σε όλες τις λειτουργίες διαχείρισης
  version: 1.0.0
  contact:
    name: softeng25-02 Team
servers:
  - url: http://localhost:8080
    description: ApiGateway development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token που λαμβάνεται από το signin endpoint
  schemas:
...
    Session:
      type: object
      properties:
        starttime:
          type: string
          description: "Format: YYYY-MM-DD HH:MM"
        endtime:
          type: string
        startsoc:
          type: integer
        endsoc:
          type: integer
        totalkwh:
          type: number
        kwhprice:
          type: number
        amount:
          type: number
...
paths:
...
  /api/v1/newsession:
    post:
      tags: [Sessions]
      summary: Δημιουργία νέας συνεδρίας φόρτισης
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pointid, starttime, endtime, totalkwh, kwhprice, amount]
              properties:
                pointid:
                  type: integer
                starttime:
                  type: string
                  format: date-time
                endtime:
                  type: string
                  format: date-time
                startsoc:
                  type: integer
                endsoc:
                  type: integer
                totalkwh:
                  type: number
                kwhprice:
                  type: number
                amount:
                  type: number
      responses:
        '200':
          description: Συνεδρία δημιουργήθηκε
        '400':
          description: Λείπουν απαιτούμενα πεδία ή μη έγκυρος φορτιστής
        '403':
          description: Φορτιστής σε χρήση ή εκτός λειτουργίας
  /api/v1/sessions/{id}/{from}/{to}:
    get:
      tags: [Sessions]
      summary: Λήψη συνεδριών φορτιστή
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID φορτιστή
          schema:
            type: integer
        - name: from
          in: path
          required: true
          description: Ημερομηνία έναρξης (YYYYMMDD)
          schema:
            type: string
          example: "20260101"
        - name: to
          in: path
          required: true
          description: Ημερομηνία λήξης (YYYYMMDD)
          schema:
            type: string
          example: "20260131"
      responses:
        '200':
          description: Λίστα συνεδριών
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '400':
          description: Μη έγκυρο ID ή μορφή ημερομηνίας

  # ==================== POINT STATUS ====================
...